+++ cbits/souffle/datastructure/ConcurrentFlyweight.h
@@ -39,6 +39,35 @@ public:
     using pointer = const value_type*;
     using reference = const value_type&;
 
+private:
+    // Effectively:
+    //  data slot_type = NONE | END | Idx index_type
+    //  The last two values in the domain of `index_type` are used to represent cases `NONE` and `END`
+    // TODO: strong type-def wrap this to prevent implicit conversions
+    using slot_type = index_type;
+    static constexpr slot_type NONE = std::numeric_limits<slot_type>::max();  // special case: `std::nullopt`
+    static constexpr slot_type END = NONE - 1;                                // special case: end iterator
+    static constexpr slot_type SLOT_MAX = END;  // +1 the largest non-special slot value
+
+    static_assert(std::is_same_v<slot_type, index_type>,
+            "conversion helpers assume they're the underlying type, "
+            "with the last two values reserved for special cases");
+    static_assert(std::is_unsigned_v<slot_type>);
+
+    /// Converts from index to slot.
+    static slot_type slot(const index_type I) {
+        // not expected to happen. you'll run out of memory long before.
+        assert(I < SLOT_MAX && "can't represent index in `slot_type` domain");
+        return static_cast<slot_type>(I);
+    }
+
+    /// Converts from slot to index.
+    static index_type index(const slot_type S) {
+        assert(S < SLOT_MAX && "slot is sentinal value; can't convert to index !!");
+        return static_cast<index_type>(S);
+    }
+
+public:
     /// Iterator with concurrent access to the datastructure.
     struct Iterator {
         using iterator_category = std::input_iterator_tag;
@@ -47,8 +76,6 @@ public:
         using reference = ConcurrentFlyweight::reference;
 
     private:
-        using slot_type = int64_t;
-
         const ConcurrentFlyweight* This;
 
         /// Access lane to the datastructure.
@@ -61,41 +88,23 @@ public:
         slot_type NextMaybeUnassignedSlot;
 
         /// Handle that owns the next slot that might be unassigned.
-        int64_t NextMaybeUnassignedHandle;
-
-        static constexpr int64_t End = std::numeric_limits<slot_type>::max();
-        static constexpr int64_t None = -1;
-
-        /// Converts from index to slot.
-        static slot_type slot(const index_type I) {
-            assert(I >= 0 && I <= std::numeric_limits<slot_type>::max());
-            return static_cast<int64_t>(I);
-        }
-
-        /// Converts from slot to index.
-        static index_type index(const slot_type S) {
-            assert(S >= 0 && S <= std::numeric_limits<index_type>::max());
-            return static_cast<index_type>(S);
-        }
+        slot_type NextMaybeUnassignedHandle = NONE;
 
     public:
         // The 'begin' iterator
         Iterator(const ConcurrentFlyweight* This, const lane_id H)
-                : This(This), Lane(H), Slot(None), NextMaybeUnassignedSlot(0),
-                  NextMaybeUnassignedHandle(None) {
+                : This(This), Lane(H), Slot(NONE), NextMaybeUnassignedSlot(0) {
             FindNextMaybeUnassignedSlot();
             MoveToNextAssignedSlot();
         }
 
         // The 'end' iterator
         Iterator(const ConcurrentFlyweight* This)
-                : This(This), Lane(0), Slot(End), NextMaybeUnassignedSlot(End),
-                  NextMaybeUnassignedHandle(None) {}
+                : This(This), Lane(0), Slot(END), NextMaybeUnassignedSlot(END) {}
 
         // The iterator starting at slot I, using access lane H.
         Iterator(const ConcurrentFlyweight* This, const lane_id H, const index_type I)
-                : This(This), Lane(H), Slot(slot(I)), NextMaybeUnassignedSlot(slot(I)),
-                  NextMaybeUnassignedHandle(None) {
+                : This(This), Lane(H), Slot(slot(I)), NextMaybeUnassignedSlot(slot(I)) {
             FindNextMaybeUnassignedSlot();
             MoveToNextAssignedSlot();
         }
@@ -148,7 +157,7 @@ public:
         }
 
         bool operator==(const Iterator& That) const {
-            return (&This == &That.This) && (Slot == That.Slot);
+            return (This == That.This) && (Slot == That.Slot);
         }
 
         bool operator!=(const Iterator& That) const {
@@ -158,45 +167,48 @@ public:
     private:
         /** Find next slot after Slot that is maybe unassigned. */
         void FindNextMaybeUnassignedSlot() {
-            NextMaybeUnassignedSlot = End;
+            NextMaybeUnassignedSlot = END;
             for (lane_id I = 0; I < This->Lanes.lanes(); ++I) {
                 const auto Lane = This->Lanes.guard(I);
-                if (This->Handles[I].NextSlot > Slot && This->Handles[I].NextSlot < NextMaybeUnassignedSlot) {
+                if ((Slot == NONE || This->Handles[I].NextSlot > Slot) &&
+                        This->Handles[I].NextSlot < NextMaybeUnassignedSlot) {
                     NextMaybeUnassignedSlot = This->Handles[I].NextSlot;
                     NextMaybeUnassignedHandle = I;
                 }
             }
-            if (NextMaybeUnassignedSlot == End) {
-                NextMaybeUnassignedSlot = This->NextSlot;
-                NextMaybeUnassignedHandle = None;
+            if (NextMaybeUnassignedSlot == END) {
+                NextMaybeUnassignedSlot = This->NextSlot.load(std::memory_order_acquire);
+                NextMaybeUnassignedHandle = NONE;
             }
         }
 
         /**
          * Move Slot to next assigned slot and return true.
-         * Otherwise the end is reached and Slot is assigned int64_t::max and return false.
+         * Otherwise the end is reached and Slot is assigned `END` and return false.
          */
         bool MoveToNextAssignedSlot() {
-            while (Slot != End) {
+            static_assert(NONE == std::numeric_limits<slot_type>::max(),
+                    "required for wrap around to 0 for begin-iterator-scan");
+            static_assert(NONE + 1 == 0, "required for wrap around to 0 for begin-iterator-scan");
+            while (Slot != END) {
+                assert(Slot + 1 < SLOT_MAX);
                 if (Slot + 1 < NextMaybeUnassignedSlot) {  // next unassigned slot not reached
                     Slot = Slot + 1;
                     return true;
                 }
 
-                if (NextMaybeUnassignedHandle == None) {  // reaching end
-                    Slot = End;
-                    NextMaybeUnassignedSlot = End;
-                    NextMaybeUnassignedHandle = None;
+                if (NextMaybeUnassignedHandle == NONE) {  // reaching end
+                    Slot = END;
+                    NextMaybeUnassignedSlot = END;
+                    NextMaybeUnassignedHandle = NONE;
                     return false;
                 }
 
-                if (NextMaybeUnassignedHandle != None) {  // maybe reaching the next unassigned slot
+                if (NextMaybeUnassignedHandle != NONE) {  // maybe reaching the next unassigned slot
                     This->Lanes.lock(NextMaybeUnassignedHandle);
                     const bool IsAssigned = (Slot + 1 < This->Handles[NextMaybeUnassignedHandle].NextSlot);
                     This->Lanes.unlock(NextMaybeUnassignedHandle);
-                    if (IsAssigned) {
-                        Slot = Slot + 1;
-                    }
+                    Slot = Slot + 1;
                     FindNextMaybeUnassignedSlot();
                     if (IsAssigned) {
                         return true;
@@ -218,7 +230,7 @@ public:
         Slots = std::make_unique<const value_type*[]>(InitialCapacity);
         Handles = std::make_unique<Handle[]>(HandleCount);
         NextSlot = (ReserveFirst ? 1 : 0);
-        MaxSlotBeforeGrow = InitialCapacity - 1;
+        SlotCount = InitialCapacity;
     }
 
     /// Initialize the datastructure with a capacity of 8 elements.
@@ -275,6 +287,7 @@ public:
     /// Assumption: the index is mapped in the datastructure.
     const Key& fetch(const lane_id H, const index_type Idx) const {
         const auto Lane = Lanes.guard(H);
+        assert(Idx < SlotCount.load(std::memory_order_relaxed));
         return Slots[Idx]->first;
     }
 
@@ -285,35 +298,61 @@ public:
     template <class... Args>
     std::pair<index_type, bool> findOrInsert(const lane_id H, Args&&... Xs) {
         const auto Lane = Lanes.guard(H);
-        int64_t Slot = Handles[H].NextSlot;
         node_type Node;
 
-        if (Slot == -1) {
-            // reserve a slot in the index, be it for now or later usage.
-            Slot = NextSlot++;
-            Node = Mapping.node(static_cast<index_type>(Slot));
+        slot_type Slot = Handles[H].NextSlot;
 
-            Handles[H].NextSlot = Slot;
-            Handles[H].NextNode = Node;
-
-            if (Slot > MaxSlotBeforeGrow) {
-                tryGrow(H);
+        // Getting the next insertion slot for the current lane may require
+        // more than one attempts if the datastructure must grow and other
+        // threads are waiting for the same lane @p H.
+        while (true) {
+            if (Slot == NONE) {
+                // Reserve a slot for the lane, the datastructure might need to
+                // grow before the slot memory location becomes available.
+                Slot = NextSlot++;
+                Handles[H].NextSlot = Slot;
+                Handles[H].NextNode = Mapping.node(static_cast<index_type>(Slot));
             }
-        } else {
-            Node = Handles[H].NextNode;
+
+            if (Slot >= SlotCount.load(std::memory_order_relaxed)) {
+                // The slot memory location is not yet available, try to
+                // grow the datastructure. Other threads in other lanes might
+                // be attempting to grow the datastructure concurrently.
+                //
+                // Anyway when this call returns the Slot memory location is
+                // available.
+                tryGrow(H);
+
+                // Reload the Slot for the current lane since another thread
+                // using the same lane may take-over the lane during tryGrow()
+                // and consume the slot before the current thread is
+                // rescheduled on the lane.
+                Slot = Handles[H].NextSlot;
+            } else {
+                // From here the slot is known, allocated and available.
+                break;
+            }
         }
 
-        // insert key in the index in advance
+        Node = Handles[H].NextNode;
+
+        // Insert key in the index in advance.
         Slots[Slot] = &Node->value();
 
         auto Res = Mapping.get(H, Node, std::forward<Args>(Xs)...);
         if (Res.second) {
-            // inserted by self
-            Handles[H].NextSlot = -1;
-            Handles[H].NextNode = node_type{};
+            // Inserted by self, slot is consumed, clear the lane's state.
+            Handles[H].clear();
             return std::make_pair(static_cast<index_type>(Slot), true);
         } else {
-            // inserted concurrently by another handle,
+            // Inserted concurrently by another thread, clearing the slot is
+            // not strictly needed but it avoids leaving a dangling pointer
+            // there.
+            //
+            // The reserved slot and node remains in the lane state so that
+            // they can be consumed by the next insertion operation on this
+            // lane.
+            Slots[Slot] = nullptr;
             return std::make_pair(Res.first->second, false);
         }
     }
@@ -323,8 +362,12 @@ private:
     using node_type = typename map_type::node_type;
 
     struct Handle {
-        /// Slot where this handle will store its next value
-        int64_t NextSlot = -1;
+        void clear() {
+            NextSlot = NONE;
+            NextNode = nullptr;
+        }
+
+        slot_type NextSlot = NONE;
         node_type NextNode = nullptr;
     };
 
@@ -346,15 +389,23 @@ private:
     map_type Mapping;
 
     // Next available slot.
-    std::atomic<std::int64_t> NextSlot;
+    std::atomic<slot_type> NextSlot;
 
-    // Maximum allowed slot index before growing
-    std::int64_t MaxSlotBeforeGrow;
+    // Number of slots.
+    std::atomic<slot_type> SlotCount;
 
+    /// Grow the datastructure if needed.
     bool tryGrow(const lane_id H) {
+        // This call may release and re-acquire the lane to
+        // allow progress of a concurrent growing operation.
+        //
+        // It is possible that another thread is waiting to
+        // enter the same lane, and that other thread might
+        // take and leave the lane before the current thread
+        // re-acquires it.
         Lanes.beforeLockAllBut(H);
 
-        if (NextSlot <= MaxSlotBeforeGrow) {
+        if (NextSlot < SlotCount) {
             // Current size is fine
             Lanes.beforeUnlockAllBut(H);
             return false;
@@ -363,12 +414,15 @@ private:
         Lanes.lockAllBut(H);
 
         {  // safe section
-            const std::size_t CurrentSize = MaxSlotBeforeGrow + 1;
-            const std::size_t NewSize = (CurrentSize << 1);  // double size policy
+            const std::size_t CurrentSize = SlotCount;
+            std::size_t NewSize = (CurrentSize << 1);  // double size policy
+            while (NewSize < NextSlot) {
+                NewSize <<= 1;  // double size
+            }
             std::unique_ptr<const value_type*[]> NewSlots = std::make_unique<const value_type*[]>(NewSize);
             std::memcpy(NewSlots.get(), Slots.get(), sizeof(const value_type*) * CurrentSize);
             Slots = std::move(NewSlots);
-            MaxSlotBeforeGrow = NewSize - 1;
+            SlotCount = NewSize;
         }
 
         Lanes.beforeUnlockAllBut(H);
@@ -393,8 +447,6 @@ public:
             const bool ReserveFirst = false, const Hash& hash = Hash(),
             const KeyEqual& key_equal = KeyEqual(), const KeyFactory& key_factory = KeyFactory())
             : Base(LaneCount, InitialCapacity, ReserveFirst, hash, key_equal, key_factory) {}
-
-    ~OmpFlyweight() {}
 
     iterator begin() const {
         return Base::begin(Base::Lanes.threadLane());
@@ -438,8 +490,6 @@ public:
             const bool ReserveFirst = false, const Hash& hash = Hash(),
             const KeyEqual& key_equal = KeyEqual(), const KeyFactory& key_factory = KeyFactory())
             : Base(NumLanes, InitialCapacity, ReserveFirst, hash, key_equal, key_factory) {}
-
-    ~SeqFlyweight() {}
 
     iterator begin() const {
         return Base::begin(0);